Moved toInputStream to JavajInputOutputStream
+ some additional changes
