+++ io/src/test/scala/fs2/io/JavaInputOutputStreamSpec.scala
@@ -7,7 +7,7 @@ import scala.annotation.tailrec
 
 
 
-class ToInputStreamSpec extends Fs2Spec {
+class JavaInputOutputStreamSpec extends Fs2Spec {
 
   "ToInputStream" - {
